cpu: eltwise: fix for issue when exp(x) is near to max value
x*log2e can exceed 128 when exp is near to max value for
polynomial with arg reduction from [-log(2)/2; log(2)/2].
This number is used in next computations to count 2^n,
where n can be 128 as it was presented earlier. 2^128 is not
representable in fp32, to around this problem 2^128 can be
divided into two numbers 2 and 2^127 which are
representable in fp32. This change fix issues in all algs
which have in their formula exp(x) i.e. gelu_erf, elu,
soft_relu, logistic, exp.